requirements.txt
streamlit==1.27.2
tensorflow-cpu==2.10.1
opencv-python-headless==4.7.0.72
numpy==1.23.5
Pillow==9.5.0
protobuf==3.20.3